move "constructQueryString" to its own component
