+++ packages/xo-web/src/common/smart-backup/preview.js
@@ -6,12 +6,12 @@ import { createSelector } from 'reselect'
 import { filter, map, pickBy } from 'lodash'
 
 import Component from './../base-component'
+import constructQueryString from '../construct-query-string'
 import Icon from './../icon'
 import Link from './../link'
 import renderXoItem from './../render-xo-item'
 import Tooltip from './../tooltip'
 import { Card, CardBlock, CardHeader } from './../card'
-import { constructQueryString } from './index'
 
 const SAMPLE_SIZE_OF_MATCHING_VMS = 3
