+++ packages/xo-web/src/xo-app/backup-ng/index.js
@@ -3,6 +3,7 @@ import ActionButton from 'action-button'
 import addSubscriptions from 'add-subscriptions'
 import Button from 'button'
 import ButtonLink from 'button-link'
+import constructQueryString from 'construct-query-string'
 import Copiable from 'copiable'
 import CopyToClipboard from 'react-copy-to-clipboard'
 import decorate from 'apply-decorators'
@@ -15,7 +16,6 @@ import Tooltip from 'tooltip'
 import { adminOnly, connectStore, routes } from 'utils'
 import { Card, CardHeader, CardBlock } from 'card'
 import { confirm } from 'modal'
-import { constructQueryString } from 'smart-backup'
 import { Container, Row, Col } from 'grid'
 import { createGetLoneSnapshots, createSelector } from 'selectors'
 import { get } from '@xen-orchestra/defined'
