+++ packages/xo-web/src/xo-app/backup/overview/index.js
@@ -2,6 +2,7 @@ import _ from 'intl'
 import ActionRowButton from 'action-row-button'
 import ButtonGroup from 'button-group'
 import Component from 'base-component'
+import constructQueryString from 'construct-query-string'
 import Icon from 'icon'
 import Link from 'link'
 import LogList from '../../logs'
@@ -13,7 +14,6 @@ import StateButton from 'state-button'
 import Tooltip from 'tooltip'
 import { confirm } from 'modal'
 import { addSubscriptions } from 'utils'
-import { constructQueryString } from 'smart-backup'
 import { createSelector } from 'selectors'
 import { Card, CardHeader, CardBlock } from 'card'
 import { filter, find, forEach, get, keyBy, map, orderBy } from 'lodash'
